+++ src/term.h
@@ -641,7 +641,7 @@ extern void term_resize(int, int);
 extern void term_scroll(int, int);
 extern void term_reset(bool full);
 extern void term_clear_scrollback(void);
-extern void term_mouse_click(mouse_button, mod_keys, pos, int count);
+extern bool term_mouse_click(mouse_button, mod_keys, pos, int count);
 extern void term_mouse_release(mouse_button, mod_keys, pos);
 extern void term_mouse_move(mod_keys, pos);
 extern void term_mouse_wheel(bool horizontal, int delta, int lines_per_notch, mod_keys, pos);